some safe typecasting

+++ source/lilc_matrix_ildl.h
@@ -16,7 +16,7 @@ void lilc_matrix<el_type> :: ildl(lilc_matrix<el_type>& L, block_diag_matrix<el_
 	
 	int lfil;
 	if (fill_factor > 1e4) lfil = ncols; //just incase users decide to enter a giant fill factor for fun...
-	else lfil = 2*fill_factor*nnz()/ncols; //roughly a factor of 2 since only lower tri. of A is stored
+	else lfil = (int) 2*fill_factor*nnz()/ncols; //roughly a factor of 2 since only lower tri. of A is stored
 	
 	const el_type alpha = (1.0+sqrt(17.0))/8.0;  //for use in pivoting.
 	el_type w1(-1), wr(-1), d1(-1), dr(-1);		//for use in bk-pivoting
